requirements.txt
anthropic>=0.45.0
google-auth>=2.20.0
requests>=2.31.0
PyJWT>=2.8.0
cryptography>=41.0.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
